Валидируем пропсы
