--- src/components/app.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-import Main from "../components/main.jsx";
-
-const App = (props) => {
-  // eslint-disable-next-line react/prop-types
-  const {offersCount} = props;
-
-  return (
-    <Main offersCount = {offersCount}/>
-  );
-};
-
-export default App;
