+++ src/index.js
@@ -1,6 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./components/app.jsx";
+import App from "./components/app/app.jsx";
+import {OFFERS} from "./const.js";
 
 const Settings = {
   OFFERS_COUNT: 312
@@ -9,6 +10,7 @@ const Settings = {
 ReactDOM.render(
     <App
       offersCount = {Settings.OFFERS_COUNT}
+      offers = {OFFERS}
     />,
     document.querySelector(`#root`)
 );
